requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Manipulation
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
Pillow>=9.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars (nice to have)
tqdm>=4.65.0

# For notebooks (if running locally)
jupyter>=1.0.0
ipykernel>=6.22.0

# Metrics and evaluation
scikit-learn>=1.2.0
